+++ docs/src/assets/menu.js
@@ -302,6 +302,10 @@ module.exports = [
       {
         name: 'Welcome',
         path: '/components/welcome'
+      },
+      {
+        name: 'WhatsappLink',
+        path: '/components/whatsapp-link'
       }
     ]
   },